feat: Create DesktopUi and Move TitleBar to kmpuiviews
This commit creates a `DesktopUi` in the `kmpuiviews` module to handle the common UI elements for desktop applications. It also moves the `CustomTitleBar` to `kmpuiviews` for better reusability across projects.

-   **DesktopUi:**
    - Created a `BaseDesktopUi` composable in `kmpuiviews` to handle window setup for desktop applications.
    - Moved `CustomTitleBar` composable to `kmpuiviews` for shared usage.
    - Updated `main.kt` in `mangaworld` to use the new `BaseDesktopUi` and removed the old title bar logic.
- **CustomTitleBar:**
+   - Updated `main.kt` to comment out a `TODO` about creating a module in kmpuiviews.
+   - Updated `DesktopUi.kt` to add comments to the `TODO` UI section.
+   - Changed `FrameWindowScope` to `ApplicationScope` in the `BaseDesktopUi` function
+**CustomTitleBar:**
     - Moved the `CustomTitleBar` composable function to the `kmpuiviews` module.
     - Updated `CustomTitleBar` to have a title, minimize click, and close click actions.
     - Updated `CustomTitleBar` to use the new `WindowDraggableArea` to allow dragging.
- - Updated `main.kt` to remove the old logic.
- - Updated `main.kt` to comment out a `TODO` about creating a module un kmpuiviews.
- - Updated `DesktopUi.kt` to add comments to the `TODO` UI section.
- - Updated `DesktopUi.kt` to use `ApplicationScope` instead of `FrameWindowScope`.
     - Updated `CustomTitleBar` to use `TopAppBar` and `IconButton`.
     - Updated imports to be kmp aware.
     - Added `PointerIcon` for Hover to the icons.
- - Updated `main.kt` to not call the old `CustomTitleBar`
+    - Updated `main.kt` to not call the old `CustomTitleBar`.
